Added edit-workflows for test items: edit-interaction.wf
This change eases tailorization of actions of editing workflow
and makes it possible to address all types of testitems via
a single term

   -form en:edit-interaction.wf

instead of having to enumerate the relevant types explicitly as
it was before

   -form en:TestItemText.form|en:TestItemShortText.form|en:TestItemMC.form|en:TestItemSC.form|...

For backward compatibility, the old enumeration is kept for a while. New test
items should be created over the new workflow, as e.g. in

   {entry -name New.Item.TextInteraction -form en:edit-interaction.wf -query p.item_type=Text}