Added substitution values for short text answers
This change adds the possibility to provide randomized substitution values to short
text questions via value sets.

Value sets are a means for a content developer to provide multiple matching answers which are inserted into the text before an exercise is shown to the end user. One can e.g. provide for a calculation exercise several input and some output values, such that the students get different calculation exercises provided. These values can also be used for the correct-when clauses.

The content developer can use percent-code delimited elements when defining the exercise:

---
Assume, you want to download a %x.what% with the size of %x.size% over a %x.type% connection with a rate of %x.rate% from %univ%.
---
and also in "correct when"
---
%x.secs%
---

the value sets can be provided via an extra field for the short-text questions and have the form
of a Tcl dict:

---
univ {WU-Vienna TU-Vienna "University of Vienna"}
x {
  {type "ADSL" rate "256 kbit/s" size "235 MB"  secs 5300 what "Powerpoint file"}
  {type "ADSL" rate "512 kbit/s" size "5.6 MB"   secs 91     what "PDF file"}
  {type "4G"      rate "80 Mbit/s"  size "270 MB" secs 27      what "PDF file"}
  {type "4G"      rate "40 Mbit/s"  size "650 MB" secs 32      what "Lecturecast Video" }
  {type "5G"      rate "1 Gbit/s"    size "520 MB" secs 4        what "Powerpoint file" }
  {type "5G"      rate "1 Gbit/s"    size "1.5 GB"  secs 12      what "Lecturecast Video" }
}
---

In this example, every student will get a randomly chosen value for the university (%univ%)
and matching elements containing the answer (e.g. download time of "270 MB" over "80 Mbit/s" is 27 seconds).
The download time is used in the correct when part, such that auto-correction can be applied.

When a student answers this exercise, the system provides random choices that are substituted in the text.
For every variable ("univ", "x", ..) different random values are used for the student.
Certainly, for other students, other numbers and results will be provided.
Note, that this value sets can be used for numeric an non-numeric exercises.

Current limitations:
- only defined for short-text questions (can be in general also for other question types)
- no elaborate user interface for entering value sets (or a thorough validator) is provided.