Handling of mutual overwrites in answer workflows
Mutual overwrites occur in answer workflows when a user manages to
open multiple browswer instances/or tabs on the same exam.

In case there was an mutual overwrite, the position as provided by the
instance attributes might deviate from the position, based on which
the actual form data was generated. So, for validating and updating
one has to change the position to the one from the form data (when
this differs). Note that the randomizer depends on property
"position" as well.

The new version avoids that the user might accidencially overwrite his
data and closes on mutual overwrite automatically the older instance
window.

Bumped version number to 5.10.0d38