Maintain revision date for "update_attribute_from_slot"
This is not a full-featured version, since it does not update the
modifying user and the modifying IP (since this data is not
always available). But this version is at least good enough
for experimenting and checking for unwanted side-effects.

xotcl-core: upadated version number to 5.10.0d40
xowf: upadated version number to 5.10.0d39