Initial support for composite questions
A composite question is a test-item containing other regular test
items. This makes it possible to define a larger test item case for
which arbitrary other questions can be associated (e.g. short text
question plus an MC and ordering question etc.). The randomization
features (shuffling, x out of N, etc.) of the included test-items are
preserved such that it is possible to present the student e.g. a sub
question where only 2 out of 5 possible alternatives are displayed.

The provided implementation builds for a composite question consisting
of a constant exam intro text and the selected sub items including
points, minutes etc. The points and minutes of the composite question
are computed as the sum of the points and minutes of the
sub-items. The actual sub-items are constructed at runtime such that
per-user randomization can be applied.

Current shortcoming: question with percent-substitutions are not
supported (since the exam intro text is placed at composite form
creation time already into the form, therefore, it is the same for all
students). Furthermore, the substitution semantics are not clear
concerning the included content.