Added support for tailored messages when autosave is rejected
Optionally, the AJAX call from autosave can be answered with
a JSON structure containing a "feedback". If this is provided, this
is presented to the user. The mechanism can be extended in the
future to include some reason code, etc. for further automatic
processing in JavaScript.

With this change, the inclass-exam-answer workflow will use this
to comminicate the reasone for rejected autosave operations in
situations, where the examtime is up.