Rename domNode argument to dom_node
tDOM defines a ::domNode command in the global namespace, which might potentially conflict with the argument name. Also, in certain configurations the api-doc machinery will treat the argument as a command usage and try to enforce e.g. the naming convention on it in acs-tcl.naming__proc_naming automated test.