Added question_ids to "details" in grading info and to the "__results" property
All of test-items is name-based to support input/export from different
OpenACS instances. In some situations, it is still necessary to pinpoint
the origin of questions, e.g. when importing pool-questions to an external
gradebook, where e.g. one wants to associate a replacement questions with
the original pool question.

Now we keep an additional question id, which will point to the pool question
or the simple question together with the question names. The instance
attribute "__results" is set once the exam-protocol was rendered
completely.

__results {
   UID1 {
    QN1 {achieved 4 achievable 6 question_id 187610}
    QN2 {achieved 0.0 achievable 1 question_id 163716}
    QN3 (achieved 0.0 achievable 2 question_id 163714}
   }
   UID2 ...
}