New feature for shared workflow definitions: define workflow specific methods
Workflows support since a long time "object-specific" code, which is code
to be executed, when a workflow instance is initialized. Now, in symmetry
to this, one can define "wf-specific" code (e.g. registering methods to
be called directly on the workflow object.

Similar to the call of ":object-specific {...some code...}", one define
additionally ":wf-specific {...some code...}". This comes handy
when e.g. definiing methods directly executed on workflowss, such as e.g.
".../someworkflow.wf?m=summary".

The methods are defined on the WorkflowContainer class, which is only available
on for shared workflows.