Added improved support for viewing and downloading of exam results
When results are available, "Points and Grades" can be used.
Results are available,
- for autocorrected questions, when the exam protocol was rendered, or
- for all types of questions, when lecturers have added manual gradings/comments.

Depending on whether a grading scheme was defined for the exam,
it is possible to view/download grades and grade statistics, or to these
just with points. It is also possible to get the total results per student
(including grades) or results per question.

The results are available
- in tabular form, with the ability to sort on the columns,
- as charts, and
- as csv-files (for loading into spreadsheets).

The version number was bumped to 5.10.1d50